[maps] fix geo line source not loaded unless maps application is opened (#159432)
closes https://github.com/elastic/kibana/issues/159408

PR consolidates source registry into a single file to ensure that all
sources are registered when only map embeddable is loaded. To prevent
regression, unit test added to ensure that all SOURCE_TYPES enum values
are contained in registry.

---------
