+++ x-pack/plugins/maps/public/connected_components/edit_layer_panel/edit_layer_panel.test.tsx
@@ -43,6 +43,13 @@ jest.mock('../../kibana_services', () => {
     getCore() {
       return {};
     },
+    getEMSSettings() {
+      return {
+        isEMSUrlSet() {
+          return false;
+        },
+      };
+    },
   };
 });
